refactor: extract DeleteModelModal component
