+++ includes/settings/js/src/components/EditModelModal.jsx
@@ -4,7 +4,7 @@ import Modal from "react-modal";
 import { ModelsContext } from "../ModelsContext";
 import Icon from "../../../../components/icons";
 import IconPicker from "./IconPicker";
-import { sprintf, __ } from "@wordpress/i18n";
+import { __ } from "@wordpress/i18n";
 
 const $ = window.jQuery;
 const { apiFetch } = wp;
@@ -89,6 +89,10 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 			boxSizing: "border-box",
 		},
 	};
+
+	useEffect(() => {
+		Modal.setAppElement("#root");
+	}, []);
 
 	return (
 		<Modal
